requirements.txt
absl-py==2.1.0
albumentations==1.4.10
einops==0.6.1
h5py==3.8.0
huggingface-hub==0.23.4
matplotlib==3.9.0
numpy==1.26.4
opencv-python==4.9.0.80
Pillow==9.4.0
protobuf==4.25.3
scikit-image==0.22.0
scikit-learn==1.5.0
scipy==1.10.1
tensorboard==2.17.0
timm==0.9.16
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
transformers==4.42.4